requirements.txt
Django==1.10.6
django-extensions==1.7.5
django-braces==1.10.0
ipython==5.1.0
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==3.2.2
psycopg2
django-celery-beat==1.1.1
django-debug-toolbar==1.6
django-model-utils==2.6
django-tables2==1.2.9
django-crispy-forms==1.6.1
django-autocomplete-light==3.2.1
factory-boy==2.8.1
django-modeladmin-reorder==0.2
xlrd==1.0.0
python-dateutil==2.6.0
python-memcached==1.59
prettytable==0.7.2
dbf==0.96.8
celery==4.2.1
flower==0.9.1
ipdb==0.10.2
xhtml2pdf==0.2b1
XlsxWriter==0.9.6
django-smuggler==0.9.0
django-dbbackup==3.2.0
dj-database-url-0.5.0
django-heroku-0.3.1
